@@ -1,8 +1,10 @@
 /** @type {import('next').NextConfig} */
+
+
 const nextConfig = {
   experimental: {
     appDir: true,
-  },
+  }
 }
 
-module.exports = nextConfig
+export default nextConfig
