@@ -1,7 +1,8 @@
 import './globals.css'
-
+import AuthContextProvider from '@/context/authcontext'
+import ExperienceContextProvider from '@/context/expContext'
 export const metadata = {
-  title: 'Create Next App',
+  title: 'Resume Revamp',
   description: 'Generated by create next app',
 }
 
@@ -12,7 +13,13 @@ export default function RootLayout({
 }) {
   return (
     <html lang="en">
-      <body>{children}</body>
+      <body className="font-exo">
+        <AuthContextProvider>
+          <ExperienceContextProvider>
+            {children}
+          </ExperienceContextProvider>
+        </AuthContextProvider>
+      </body>
     </html>
   )
 }
